issue #8: refactor db functions location
